refactor!: Moved `Item`s into separate packages
BREAKIING CHANGES:
- Existing Imports of most items except `ActionItem` and `Item` will break. It is now under `input` and `display` packages

What's New:
- `input` and `display` packages to reflect the role of the `Item` subclasses
