+++ lib/src/main/java/examples/StressTest.java
@@ -2,7 +2,7 @@ package examples;
 
 import io.github.bfur64.menu.MenuManager;
 import io.github.bfur64.menu.item.ActionItem;
-import io.github.bfur64.menu.item.StaticText;
+import io.github.bfur64.menu.item.display.StaticText;
 import io.github.bfur64.terminal.Terminal;
 import io.github.bfur64.terminal.interfaces.TerminalBackend;
