+++ lib/src/main/java/io/github/bfur64/menu/item/display/LineBreak.java
@@ -1,4 +1,4 @@
-package io.github.bfur64.menu.item;
+package io.github.bfur64.menu.item.display;
 
 import org.jspecify.annotations.NullMarked;
