+++ lib/src/main/java/io/github/bfur64/menu/item/input/InputItem.java
@@ -1,6 +1,7 @@
-package io.github.bfur64.menu.item;
+package io.github.bfur64.menu.item.input;
 
 import io.github.bfur64.menu.input.InputHandler;
+import io.github.bfur64.menu.item.Item;
 import io.github.bfur64.menu.utils.Property;
 import io.github.bfur64.terminal.input.KeyStroke;
 import io.github.bfur64.terminal.input.KeyType;
