+++ lib/src/main/java/io/github/bfur64/menu/item/input/KeyInputItem.java
@@ -1,4 +1,4 @@
-package io.github.bfur64.menu.item;
+package io.github.bfur64.menu.item.input;
 
 import io.github.bfur64.menu.utils.Property;
 import io.github.bfur64.terminal.input.KeyStroke;
